feat: destructuring form input and buttons on the login page into components
